fix: refactor visual test scripts
